1단계 : Auth Package 분리 및 Auth -> nextstep 의존성 제거
