+++ src/main/java/auth/AdminInterceptor.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package auth;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.servlet.HandlerInterceptor;
