+++ src/main/java/auth/AuthBeanConfiguration.java
@@ -1,0 +1,13 @@
+package auth;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AuthBeanConfiguration {
+
+  @Bean
+  public JwtTokenProvider jwtTokenProvider() {
+    return new JwtTokenProvider();
+  }
+}
