+++ src/main/java/auth/AuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package auth;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
