+++ src/main/java/auth/JwtTokenProvider.java
@@ -1,12 +1,10 @@
-package nextstep.auth;
+package auth;
 
 import io.jsonwebtoken.*;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 
 import java.util.Date;
 
-@Component
 public class JwtTokenProvider {
     @Value("${security.jwt.token.secret-key}")
     private String secretKey;
