+++ src/main/java/auth/LoginController.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package auth;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
