+++ src/main/java/auth/LoginMember.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
