+++ src/main/java/auth/LoginMemberArgumentResolver.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package auth;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.http.HttpHeaders;
