+++ src/main/java/auth/MemberDetail.java
@@ -1,0 +1,5 @@
+package auth;
+
+public interface MemberDetail {
+
+}
