+++ src/main/java/auth/TokenRequest.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package auth;
 
 public class TokenRequest {
     private String username;
